backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Andrada Tinta"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "chromadb (>=1.0.20,<2.0.0)",
    "openai (>=1.101.0,<2.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[project.scripts]
start = "uvicorn.backend.api:app"